moved functionality to dashboards
